--- src/main/java/Moves/Empty.java
@@ -1,12 +1,0 @@
-package Moves;
-
-public class Empty implements Move {
-    @Override
-    public String toString() {
-        return "EMPTY";
-    }
-
-    public String pretty(){
-        return "-";
-    }
-}
