--- src/main/java/Moves/GiveUp.java
@@ -1,12 +1,0 @@
-package Moves;
-
-public class GiveUp implements Move {
-    @Override
-    public String toString() {
-        return "GIVEUP";
-    }
-
-    public String pretty(){
-        return "Poddał się";
-    }
-}
