--- src/main/java/Moves/Pass.java
@@ -1,12 +1,0 @@
-package Moves;
-
-public class Pass implements Move {
-    @Override
-    public String toString() {
-        return "PASS";
-    }
-
-    public String pretty(){
-        return "Pass";
-    }
-}
