--- src/main/java/Server/Exceprions/ConnectionTroubleException.java
@@ -1,5 +1,0 @@
-package Server.Exceprions;
-
-public class ConnectionTroubleException extends Exception {
-}
-
