--- src/main/java/Server/Exceprions/GiveUpException.java
@@ -1,4 +1,0 @@
-package Server.Exceprions;
-
-public class GiveUpException extends Exception{
-}
