--- src/main/java/Server/Exceprions/IllegalMoveException.java
@@ -1,5 +1,0 @@
-package Server.Exceprions;
-
-public class IllegalMoveException extends Exception {
-}
-
