--- src/main/java/Server/Exceprions/IllegalPositionException.java
@@ -1,8 +1,0 @@
-package Server.Exceprions;
-
-public class IllegalPositionException extends IllegalMoveException {
-    @Override
-    public String toString() {
-        return "ILLEGALPOSITION";
-    }
-}
