--- src/main/java/Server/Exceprions/KoException.java
@@ -1,8 +1,0 @@
-package Server.Exceprions;
-
-public class KoException extends IllegalMoveException {
-    @Override
-    public String toString() {
-        return "KO";
-    }
-}
