--- src/main/java/Server/Exceprions/SuicideException.java
@@ -1,8 +1,0 @@
-package Server.Exceprions;
-
-public class SuicideException extends IllegalMoveException {
-    @Override
-    public String toString() {
-        return "SUICIDE";
-    }
-}
