+++ src/main/java/moves/Empty.java
@@ -1,0 +1,12 @@
+package moves;
+
+public class Empty implements Move {
+    @Override
+    public String toString() {
+        return "EMPTY";
+    }
+
+    public String pretty(){
+        return "-";
+    }
+}
