+++ src/main/java/moves/GiveUp.java
@@ -1,0 +1,12 @@
+package moves;
+
+public class GiveUp implements Move {
+    @Override
+    public String toString() {
+        return "GIVEUP";
+    }
+
+    public String pretty(){
+        return "Poddał się";
+    }
+}
