+++ src/main/java/moves/Pass.java
@@ -1,0 +1,12 @@
+package moves;
+
+public class Pass implements Move {
+    @Override
+    public String toString() {
+        return "PASS";
+    }
+
+    public String pretty(){
+        return "Pass";
+    }
+}
