+++ src/main/java/server/exceprions/ConnectionTroubleException.java
@@ -1,0 +1,5 @@
+package server.exceprions;
+
+public class ConnectionTroubleException extends Exception {
+}
+
