+++ src/main/java/server/exceprions/GiveUpException.java
@@ -1,0 +1,4 @@
+package server.exceprions;
+
+public class GiveUpException extends Exception{
+}
