+++ src/main/java/server/exceprions/IllegalMoveException.java
@@ -1,0 +1,5 @@
+package server.exceprions;
+
+public class IllegalMoveException extends Exception {
+}
+
