+++ src/main/java/server/exceprions/IllegalPositionException.java
@@ -1,0 +1,8 @@
+package server.exceprions;
+
+public class IllegalPositionException extends IllegalMoveException {
+    @Override
+    public String toString() {
+        return "ILLEGALPOSITION";
+    }
+}
