+++ src/main/java/server/exceprions/KoException.java
@@ -1,0 +1,8 @@
+package server.exceprions;
+
+public class KoException extends IllegalMoveException {
+    @Override
+    public String toString() {
+        return "KO";
+    }
+}
