+++ src/main/java/server/exceprions/SuicideException.java
@@ -1,0 +1,8 @@
+package server.exceprions;
+
+public class SuicideException extends IllegalMoveException {
+    @Override
+    public String toString() {
+        return "SUICIDE";
+    }
+}
